requirements.txt
requests==2.25.0
absl-py==1.0.0